refactor: Share addTemplateFilesToZip()
